Reduce FBodyInstance size down from 472 to 352. Refactor customization code for primitive component and bodysetup so that it's all in one place. This was needed to make properties show up in human order vs best packing order
#JIRA UE-14619

[CL 2622427 by Ori Cohen in Main branch]

+++ Engine/Source/Runtime/Engine/Private/PhysicsEngine/PhysScene.cpp
@@ -24,7 +24,7 @@
 #define USE_ADAPTIVE_FORCES_FOR_ASYNC_SCENE			1
 #define USE_SPECIAL_FRICTION_MODEL_FOR_ASYNC_SCENE	0
 
-static int32 PhysXSceneCount = 1;
+static int16 PhysXSceneCount = 1;
 static const int PhysXSlowRebuildRate = 10;
 
 EPhysicsSceneType FPhysScene::SceneType_AssumesLocked(const FBodyInstance* BodyInstance) const
